@@ -17,23 +17,21 @@
         }
     ],
     "require": {
-        "php": "^7.0 || ^8.0",
-        "symfony/framework-bundle": "^3.0 || ^4.0 || ^5.0",
-        "ivory/google-map": "^3.0"
+        "php": "^8.1",
+        "symfony/framework-bundle": "^6.4 || ^7.0",
+        "ivory/google-map": "^6.0.1"
     },
     "require-dev": {
-        "ivory/serializer-bundle": "^1.0",
-        "friendsofphp/php-cs-fixer": "^2.0",
-        "php-http/cache-plugin": "^1.4",
-        "php-http/guzzle6-adapter": "^1.0",
-        "phpunit/phpunit": "^6.5",
-        "phpunit/phpunit-selenium": "^4.1.0",
-        "symfony/cache": "^3.2 || ^4.0",
-        "symfony/form": "^2.7 || ^3.0 || ^4.0",
-        "symfony/phpunit-bridge": "^2.7 || ^3.0 || ^4.0",
-        "symfony/templating": "^2.7 || ^3.0 || ^4.0",
-        "symfony/yaml": "^2.7 || ^3.0 || ^4.0",
-        "twig/twig": "^2.7 || ^3.0"
+        "friendsofphp/php-cs-fixer": "^3.49",
+        "php-http/cache-plugin": "^2.0",
+        "php-http/guzzle7-adapter": "^1.0",
+        "phpunit/phpunit": "^9.6",
+        "phpunit/phpunit-selenium": "^9.0",
+        "symfony/cache": "^6.4 || ^7.0",
+        "symfony/form": "^6.4 || ^7.0",
+        "symfony/phpunit-bridge": "^6.4 || ^7.0",
+        "symfony/yaml": "^6.4 || ^7.0",
+        "twig/twig": "^3.8"
     },
     "autoload": {
         "psr-4": { "Ivory\\GoogleMapBundle\\": "" }
