@@ -20,23 +20,13 @@ use Twig\TwigFunction;
  */
 class ApiExtension extends AbstractExtension
 {
-    /**
-     * @var ApiHelper
-     */
-    private $apiHelper;
-
-    /**
-     * @param ApiHelper $apiHelper
-     */
-    public function __construct(ApiHelper $apiHelper)
-    {
-        $this->apiHelper = $apiHelper;
-    }
+    public function __construct(private ApiHelper $apiHelper)
+    {}
 
     /**
      * {@inheritdoc}
      */
-    public function getFunctions()
+    public function getFunctions(): array
     {
         $functions = [];
 
@@ -49,18 +39,13 @@ class ApiExtension extends AbstractExtension
 
     /**
      * @param object[] $objects
-     *
-     * @return string
      */
-    public function render(array $objects)
+    public function render(array $objects): string
     {
         return $this->apiHelper->render($objects);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
+    public function getName(): string
     {
         return 'ivory_google_api';
     }
@@ -68,7 +53,7 @@ class ApiExtension extends AbstractExtension
     /**
      * @return string[]
      */
-    private function getMapping()
+    private function getMapping(): array
     {
         return ['ivory_google_api' => 'render'];
     }
