@@ -21,23 +21,13 @@ use Twig\TwigFunction;
  */
 class StaticMapExtension extends AbstractExtension
 {
-    /**
-     * @var StaticMapHelper
-     */
-    private $staticMapHelper;
-
-    /**
-     * @param StaticMapHelper $staticMapHelper
-     */
-    public function __construct(StaticMapHelper $staticMapHelper)
-    {
-        $this->staticMapHelper = $staticMapHelper;
-    }
+    public function __construct(private StaticMapHelper $staticMapHelper)
+    {}
 
     /**
      * {@inheritdoc}
      */
-    public function getFunctions()
+    public function getFunctions(): array
     {
         $functions = [];
 
@@ -48,20 +38,12 @@ class StaticMapExtension extends AbstractExtension
         return $functions;
     }
 
-    /**
-     * @param Map $map
-     *
-     * @return string
-     */
-    public function render(Map $map)
+    public function render(Map $map): string
     {
         return $this->staticMapHelper->render($map);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getName()
+    public function getName(): string
     {
         return 'ivory_google_map_static';
     }
@@ -69,7 +51,7 @@ class StaticMapExtension extends AbstractExtension
     /**
      * @return string[]
      */
-    private function getMapping()
+    private function getMapping(): array
     {
         return ['ivory_google_map_static' => 'render'];
     }
