@@ -29,9 +29,9 @@ class IvoryGoogleMapExtension extends ConfigurableExtension
     /**
      * {@inheritdoc}
      */
-    protected function loadInternal(array $config, ContainerBuilder $container)
+    protected function loadInternal(array $mergedConfig, ContainerBuilder $container): void
     {
-        $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
 
         $resources = [
             'form',
@@ -40,23 +40,22 @@ class IvoryGoogleMapExtension extends ConfigurableExtension
             'helper/renderer',
             'helper/subscriber',
             'helper/utility',
-            'templating',
             'twig',
         ];
 
         foreach ($resources as $resource) {
-            $loader->load($resource.'.xml');
+            $loader->load($resource . '.xml');
         }
 
-        $this->loadMapConfig($config['map'], $container);
-        $this->loadStaticMapConfig($config['static_map'], $container);
-        $this->loadServicesConfig($config, $container, $loader);
+        $this->loadMapConfig($mergedConfig['map'], $container);
+        $this->loadStaticMapConfig($mergedConfig['static_map'], $container);
+        $this->loadServicesConfig($mergedConfig, $container, $loader);
     }
 
     /**
      * @param mixed[] $config
      */
-    private function loadMapConfig(array $config, ContainerBuilder $container)
+    private function loadMapConfig(array $config, ContainerBuilder $container): void
     {
         $container
             ->getDefinition('ivory.google_map.helper.renderer.loader')
@@ -78,7 +77,7 @@ class IvoryGoogleMapExtension extends ConfigurableExtension
     /**
      * @param mixed[] $config
      */
-    private function loadStaticMapConfig(array $config, ContainerBuilder $container)
+    private function loadStaticMapConfig(array $config, ContainerBuilder $container): void
     {
         if (isset($config['api_key'])) {
             $container
@@ -100,39 +99,23 @@ class IvoryGoogleMapExtension extends ConfigurableExtension
     /**
      * @param mixed[] $config
      */
-    private function loadServicesConfig(array $config, ContainerBuilder $container, LoaderInterface $loader)
+    private function loadServicesConfig(array $config, ContainerBuilder $container, LoaderInterface $loader): void
     {
         $services = [
-            'direction'          => true,
-            'distance_matrix'    => true,
-            'elevation'          => true,
-            'geocoder'           => true,
+            'direction' => true,
+            'distance_matrix' => true,
+            'elevation' => true,
+            'geocoder' => true,
             'place_autocomplete' => true,
-            'place_detail'       => true,
-            'place_photo'        => false,
-            'place_search'       => true,
-            'time_zone'          => true,
+            'place_detail' => true,
+            'place_photo' => false,
+            'place_search' => true,
+            'time_zone' => true,
         ];
-
-        $serializerLoaded = false;
-        $loadSerializer = function() use ($container, $loader) {
-            $loader->load('service/serializer.xml');
-
-            if ($container->hasParameter('kernel.project_dir')) {
-                $container
-                    ->getDefinition('ivory.google_map.serializer.loader')
-                    ->replaceArgument(0, '%kernel.project_dir%/vendor/ivory/google-map/src/Service/Serializer');
-            }
-        };
 
         foreach ($services as $service => $http) {
             if (!isset($config[$service])) {
                 continue;
-            }
-
-            if ($http && !$serializerLoaded) {
-                $loadSerializer();
-                $serializerLoaded = true;
             }
 
             $this->loadServiceConfig($service, $config[$service], $container, $loader, $http);
@@ -140,27 +123,25 @@ class IvoryGoogleMapExtension extends ConfigurableExtension
     }
 
     /**
-     * @param string  $service
      * @param mixed[] $config
-     * @param bool    $http
      *
      * @throws Exception
      */
     private function loadServiceConfig(
-        string $service,
-        array $config,
+        string           $service,
+        array            $config,
         ContainerBuilder $container,
-        LoaderInterface $loader,
-        $http = true
-    ) {
-        $loader->load('service/'.$service.'.xml');
-        $definition = $container->getDefinition($serviceName = 'ivory.google_map.'.$service);
+        LoaderInterface  $loader,
+                         $http = true
+    ): void
+    {
+        $loader->load('service/' . $service . '.xml');
+        $definition = $container->getDefinition($serviceName = 'ivory.google_map.' . $service);
 
         if ($http) {
             $definition
                 ->addArgument(new Reference($config['client']))
-                ->addArgument(new Reference($config['message_factory']))
-                ->addArgument(new Reference('ivory.serializer'));
+                ->addArgument(new Reference($config['message_factory']));
         }
 
         if ($http && isset($config['format'])) {
@@ -175,7 +156,7 @@ class IvoryGoogleMapExtension extends ConfigurableExtension
             $businessAccountConfig = $config['business_account'];
 
             $container->setDefinition(
-                $businessAccountName = $serviceName.'.business_account',
+                $businessAccountName = $serviceName . '.business_account',
                 new Definition(BusinessAccount::class, [
                     $businessAccountConfig['client_id'],
                     $businessAccountConfig['secret'],
