@@ -28,7 +28,7 @@ class IvoryGoogleMapBundle extends Bundle
     /**
      * {@inheritdoc}
      */
-    public function build(ContainerBuilder $container)
+    public function build(ContainerBuilder $container): void
     {
         $container
             ->addCompilerPass(new CleanTemplatingPass())
