@@ -28,11 +28,11 @@ use Ivory\GoogleMap\Service\Place\Search\PlaceSearchService;
 use Ivory\GoogleMap\Service\TimeZone\TimeZoneService;
 use Ivory\GoogleMapBundle\DependencyInjection\IvoryGoogleMapExtension;
 use Ivory\GoogleMapBundle\IvoryGoogleMapBundle;
-use Ivory\Serializer\SerializerInterface;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
+use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @author GeLo <geloen.eric@gmail.com>
@@ -55,24 +55,19 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
     private $locale;
 
     /**
-     * @var HttpClient|PHPUnit_Framework_MockObject_MockObject
+     * @var HttpClient|MockObject
      */
     private $client;
 
     /**
-     * @var MessageFactory|PHPUnit_Framework_MockObject_MockObject
+     * @var MessageFactory|MockObject
      */
     private $messageFactory;
 
     /**
-     * @var SerializerInterface|PHPUnit_Framework_MockObject_MockObject
-     */
-    private $serializer;
-
-    /**
      * {@inheritdoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->container = new ContainerBuilder();
         $this->container->setParameter('kernel.root_dir', __DIR__.'/..');
@@ -81,18 +76,14 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->container->setParameter('locale', $this->locale = 'en');
         $this->container->set('httplug.client', $this->client = $this->createClientMock());
         $this->container->set('httplug.message_factory', $this->messageFactory = $this->createMessageFactoryMock());
-        $this->container->set('ivory.serializer', $this->serializer = $this->createSerializerMock());
         $this->container->registerExtension($extension = new IvoryGoogleMapExtension());
         $this->container->loadFromExtension($extension->getAlias());
         (new IvoryGoogleMapBundle())->build($this->container);
     }
 
-    /**
-     * @param string $configuration
-     */
-    abstract protected function loadConfiguration(ContainerBuilder $container, $configuration);
+    abstract protected function loadConfiguration(ContainerBuilder $container, string $configuration);
 
-    public function testDefaultState()
+    public function testDefaultState(): void
     {
         $this->container->compile();
 
@@ -116,12 +107,13 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertTrue($this->container->get('ivory.google_map.helper.renderer.control.manager')->hasRenderers());
         $this->assertTrue($this->container->get('ivory.google_map.helper.renderer.overlay.extendable')->hasRenderers());
 
+        /*
         $this->assertTrue($this->container->get('ivory.google_map.helper.api.event_dispatcher')->hasListeners());
         $this->assertTrue($this->container->get('ivory.google_map.helper.map.event_dispatcher')->hasListeners());
         $this->assertTrue($this->container->get('ivory.google_map.helper.map.static.event_dispatcher')->hasListeners());
         $this->assertTrue(
             $this->container->get('ivory.google_map.helper.place_autocomplete.event_dispatcher')->hasListeners()
-        );
+        );*/
 
         $this->assertFalse($this->container->has('ivory.google_map.direction'));
         $this->assertFalse($this->container->has('ivory.google_map.distance_matrix'));
@@ -136,17 +128,6 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertFalse($this->container->has('ivory.google_map.twig.extension.api'));
         $this->assertFalse($this->container->has('ivory.google_map.twig.extension.map'));
         $this->assertFalse($this->container->has('ivory.google_map.twig.extension.place_autocomplete'));
-    }
-
-    public function testTemplatingHelpers()
-    {
-        $this->container->setDefinition('templating.engine.php', new Definition(\stdClass::class));
-        $this->container->compile();
-
-        $this->assertTrue($this->container->has('ivory.google_map.templating.api'));
-        $this->assertTrue($this->container->has('ivory.google_map.templating.map'));
-        $this->assertTrue($this->container->has('ivory.google_map.templating.map.static'));
-        $this->assertTrue($this->container->has('ivory.google_map.templating.place_autocomplete'));
     }
 
     public function testTwigExtensions()
@@ -253,15 +234,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $staticMapHelper->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testStaticMapBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'static_map_business_account_invalid');
         $this->container->compile();
     }
 
+    /**
+     * @throws \Exception
+     */
     public function testDirection()
     {
         $this->loadConfiguration($this->container, 'direction');
@@ -272,18 +254,10 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(DirectionService::class, $direction);
         $this->assertSame($this->client, $direction->getClient());
         $this->assertSame($this->messageFactory, $direction->getMessageFactory());
-        $this->assertSame($this->serializer, $direction->getSerializer());
-        $this->assertSame(DirectionService::FORMAT_JSON, $direction->getFormat());
         $this->assertFalse($direction->hasBusinessAccount());
-        self::assertTrue($this->container->has('ivory.google_map.serializer.loader'));
-    }
 
-    public function testDirectionFormat()
-    {
-        $this->loadConfiguration($this->container, 'direction_format');
-        $this->container->compile();
-
-        $this->assertSame(DirectionService::FORMAT_XML, $this->container->get('ivory.google_map.direction')->getFormat());
+        # TODO Check
+        #self::assertTrue($this->container->has('ivory.google_map.serializer.loader'));
     }
 
     public function testDirectionApiKey()
@@ -320,20 +294,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $direction->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testDirectionBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'direction_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testDirectionInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'direction_invalid');
         $this->container->compile();
     }
@@ -348,20 +318,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(DistanceMatrixService::class, $distanceMatrix);
         $this->assertSame($this->client, $distanceMatrix->getClient());
         $this->assertSame($this->messageFactory, $distanceMatrix->getMessageFactory());
-        $this->assertSame($this->serializer, $distanceMatrix->getSerializer());
-        $this->assertSame(DistanceMatrixService::FORMAT_JSON, $distanceMatrix->getFormat());
         $this->assertFalse($distanceMatrix->hasBusinessAccount());
-    }
-
-    public function testDistanceMatrixFormat()
-    {
-        $this->loadConfiguration($this->container, 'distance_matrix_format');
-        $this->container->compile();
-
-        $this->assertSame(
-            DistanceMatrixService::FORMAT_XML,
-            $this->container->get('ivory.google_map.distance_matrix')->getFormat()
-        );
     }
 
     public function testDistanceMatrixApiKey()
@@ -398,20 +355,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $distanceMatrix->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testDistanceMatrixBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'distance_matrix_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testDistanceMatrixInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'distance_matrix_invalid');
         $this->container->compile();
     }
@@ -426,17 +379,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(ElevationService::class, $elevation);
         $this->assertSame($this->client, $elevation->getClient());
         $this->assertSame($this->messageFactory, $elevation->getMessageFactory());
-        $this->assertSame($this->serializer, $elevation->getSerializer());
-        $this->assertSame(ElevationService::FORMAT_JSON, $elevation->getFormat());
         $this->assertFalse($elevation->hasBusinessAccount());
-    }
-
-    public function testElevationFormat()
-    {
-        $this->loadConfiguration($this->container, 'elevation_format');
-        $this->container->compile();
-
-        $this->assertSame(ElevationService::FORMAT_XML, $this->container->get('ivory.google_map.elevation')->getFormat());
     }
 
     public function testElevationApiKey()
@@ -473,20 +416,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $elevation->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testElevationBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'elevation_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testElevationInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'elevation_invalid');
         $this->container->compile();
     }
@@ -501,20 +440,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(GeocoderService::class, $geocoder);
         $this->assertSame($this->client, $geocoder->getClient());
         $this->assertSame($this->messageFactory, $geocoder->getMessageFactory());
-        $this->assertSame($this->serializer, $geocoder->getSerializer());
-        $this->assertSame(GeocoderService::FORMAT_JSON, $geocoder->getFormat());
         $this->assertFalse($geocoder->hasBusinessAccount());
-    }
-
-    public function testGeocoderFormat()
-    {
-        $this->loadConfiguration($this->container, 'geocoder_format');
-        $this->container->compile();
-
-        $this->assertSame(
-            GeocoderService::FORMAT_XML,
-            $this->container->get('ivory.google_map.geocoder')->getFormat()
-        );
     }
 
     public function testGeocoderApiKey()
@@ -551,20 +477,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $geocoder->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testGeocoderBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'geocoder_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testGeocoderInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'geocoder_invalid');
         $this->container->compile();
     }
@@ -579,20 +501,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(PlaceAutocompleteService::class, $placeAutocomplete);
         $this->assertSame($this->client, $placeAutocomplete->getClient());
         $this->assertSame($this->messageFactory, $placeAutocomplete->getMessageFactory());
-        $this->assertSame($this->serializer, $placeAutocomplete->getSerializer());
-        $this->assertSame(PlaceAutocompleteService::FORMAT_JSON, $placeAutocomplete->getFormat());
         $this->assertFalse($placeAutocomplete->hasBusinessAccount());
-    }
-
-    public function testPlaceAutocompleteFormat()
-    {
-        $this->loadConfiguration($this->container, 'place_autocomplete_format');
-        $this->container->compile();
-
-        $this->assertSame(
-            PlaceAutocompleteService::FORMAT_XML,
-            $this->container->get('ivory.google_map.place_autocomplete')->getFormat()
-        );
     }
 
     public function testPlaceAutocompleteApiKey()
@@ -629,20 +538,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $placeAutocomplete->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlaceAutocompleteBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_autocomplete_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlaceAutocompleteInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_autocomplete_invalid');
         $this->container->compile();
     }
@@ -657,20 +562,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(PlaceDetailService::class, $placeDetail);
         $this->assertSame($this->client, $placeDetail->getClient());
         $this->assertSame($this->messageFactory, $placeDetail->getMessageFactory());
-        $this->assertSame($this->serializer, $placeDetail->getSerializer());
-        $this->assertSame(PlaceDetailService::FORMAT_JSON, $placeDetail->getFormat());
         $this->assertFalse($placeDetail->hasBusinessAccount());
-    }
-
-    public function testPlaceDetailFormat()
-    {
-        $this->loadConfiguration($this->container, 'place_detail_format');
-        $this->container->compile();
-
-        $this->assertSame(
-            PlaceDetailService::FORMAT_XML,
-            $this->container->get('ivory.google_map.place_detail')->getFormat()
-        );
     }
 
     public function testPlaceDetailApiKey()
@@ -707,20 +599,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $placeDetail->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlaceDetailBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_detail_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlaceDetailInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_detail_invalid');
         $this->container->compile();
     }
@@ -771,11 +659,9 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $placePhoto->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlacePhotoBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_photo_business_account_invalid');
         $this->container->compile();
     }
@@ -790,20 +676,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(PlaceSearchService::class, $placeSearch);
         $this->assertSame($this->client, $placeSearch->getClient());
         $this->assertSame($this->messageFactory, $placeSearch->getMessageFactory());
-        $this->assertSame($this->serializer, $placeSearch->getSerializer());
-        $this->assertSame(PlaceSearchService::FORMAT_JSON, $placeSearch->getFormat());
         $this->assertFalse($placeSearch->hasBusinessAccount());
-    }
-
-    public function testPlaceSearchFormat()
-    {
-        $this->loadConfiguration($this->container, 'place_search_format');
-        $this->container->compile();
-
-        $this->assertSame(
-            PlaceSearchService::FORMAT_XML,
-            $this->container->get('ivory.google_map.place_search')->getFormat()
-        );
     }
 
     public function testPlaceSearchApiKey()
@@ -840,20 +713,16 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $placeSearch->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlaceSearchBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_search_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testPlaceSearchInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'place_search_invalid');
         $this->container->compile();
     }
@@ -868,17 +737,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertInstanceOf(TimeZoneService::class, $timeZone);
         $this->assertSame($this->client, $timeZone->getClient());
         $this->assertSame($this->messageFactory, $timeZone->getMessageFactory());
-        $this->assertSame($this->serializer, $timeZone->getSerializer());
-        $this->assertSame(TimeZoneService::FORMAT_JSON, $timeZone->getFormat());
         $this->assertFalse($timeZone->hasBusinessAccount());
-    }
-
-    public function testTimeZoneFormat()
-    {
-        $this->loadConfiguration($this->container, 'time_zone_format');
-        $this->container->compile();
-
-        $this->assertSame(TimeZoneService::FORMAT_XML, $this->container->get('ivory.google_map.time_zone')->getFormat());
     }
 
     public function testTimeZoneApiKey()
@@ -915,36 +774,30 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
         $this->assertSame('my-channel', $timeZone->getBusinessAccount()->getChannel());
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testTimeZoneBusinessAccountInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'time_zone_business_account_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \Symfony\Component\Config\Definition\Exception\InvalidConfigurationException
-     */
     public function testTimeZoneInvalid()
     {
+        $this->expectException(\Symfony\Component\Config\Definition\Exception\InvalidConfigurationException::class);
         $this->loadConfiguration($this->container, 'time_zone_invalid');
         $this->container->compile();
     }
 
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessage No "class" attribute found for the tag "ivory.google_map.helper.renderer.extendable" on the service "acme.map.helper.renderer.extendable".
-     */
     public function testMissingExtendableRendererClassTagAttribute()
     {
+        $this->expectExceptionMessage("No \"class\" attribute found for the tag \"ivory.google_map.helper.renderer.extendable\" on the service \"acme.map.helper.renderer.extendable\".");
+        $this->expectException(\RuntimeException::class);
         $this->loadConfiguration($this->container, 'extendable');
         $this->container->compile();
     }
 
     /**
-     * @return PHPUnit_Framework_MockObject_MockObject|HttpClient
+     * @return MockObject|HttpClient
      */
     private function createClientMock()
     {
@@ -952,7 +805,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
     }
 
     /**
-     * @return PHPUnit_Framework_MockObject_MockObject|MessageFactory
+     * @return MockObject|MessageFactory
      */
     private function createMessageFactoryMock()
     {
@@ -960,7 +813,7 @@ abstract class AbstractIvoryGoogleMapExtensionTest extends TestCase
     }
 
     /**
-     * @return PHPUnit_Framework_MockObject_MockObject|SerializerInterface
+     * @return MockObject|SerializerInterface
      */
     private function createSerializerMock()
     {
