@@ -23,7 +23,7 @@ class YamlIvoryGoogleMapExtensionTest extends AbstractIvoryGoogleMapExtensionTes
     /**
      * {@inheritdoc}
      */
-    protected function loadConfiguration(ContainerBuilder $container, $configuration)
+    protected function loadConfiguration(ContainerBuilder $container, string $configuration): void
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Fixtures/Config/Yaml'));
         $loader->load($configuration.'.yml');
