@@ -28,12 +28,12 @@ class IvoryGoogleMapBundleTest extends TestCase
     /**
      * {@inheritdoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->bundle = new IvoryGoogleMapBundle();
     }
 
-    public function testBundle()
+    public function testBundle(): void
     {
         $this->assertInstanceOf(Bundle::class, $this->bundle);
     }
