@@ -14,6 +14,7 @@ namespace Ivory\GoogleMapBundle\Tests\Twig;
 use Ivory\GoogleMap\Helper\PlaceAutocompleteHelper;
 use Ivory\GoogleMap\Place\Autocomplete;
 use Ivory\GoogleMapBundle\Twig\PlaceAutocompleteExtension;
+use PHPUnit\Framework\MockObject\MockObject;
 
 /**
  * @author GeLo <geloen.eric@gmail.com>
@@ -21,7 +22,7 @@ use Ivory\GoogleMapBundle\Twig\PlaceAutocompleteExtension;
 class PlaceAutocompleteExtensionTest extends AbstractExtensionTest
 {
     /**
-     * @var PlaceAutocompleteHelper|\PHPUnit_Framework_MockObject_MockObject
+     * @var PlaceAutocompleteHelper|MockObject
      */
     private $placeAutocompleteHelper;
 
@@ -75,7 +76,7 @@ class PlaceAutocompleteExtensionTest extends AbstractExtensionTest
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|PlaceAutocompleteHelper
+     * @return MockObject|PlaceAutocompleteHelper
      */
     private function createPlaceAutocompleteHelperMock()
     {
@@ -83,7 +84,7 @@ class PlaceAutocompleteExtensionTest extends AbstractExtensionTest
     }
 
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|Autocomplete
+     * @return MockObject|Autocomplete
      */
     private function createAutocompleteMock()
     {
