@@ -9,6 +9,8 @@
  * file that was distributed with this source code.
  */
 
+use PHPUnit\Extensions\Selenium2TestCase;
+
 if (isset($_SERVER['CACHE_PATH'])) {
     $_SERVER['CACHE_PATH'] = __DIR__.'/../'.$_SERVER['CACHE_PATH'];
 
@@ -19,4 +21,4 @@ if (isset($_SERVER['CACHE_PATH'])) {
 
 require_once __DIR__.'/../vendor/autoload.php';
 
-\PHPUnit_Extensions_Selenium2TestCase::shareSession(true);
+Selenium2TestCase::shareSession(true);
